requirements.txt
# Core
Django==5.0.7
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Pillow==10.4.0
stripe==8.5.0
requests==2.32.3
# Async / Tasks
celery==5.4.0
redis==5.0.7
django-redis==5.4.0
# Security / Auth
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
whitenoise==6.6.0
# Utilities
django-filter==24.3
django-environ==0.11.2
gunicorn==22.0.0
# Dev & Debug
django-debug-toolbar==4.4.6
black==24.8.0
flake8==7.1.1
django-extensions

